+++ tracer/src/Datadog.Trace/ClrProfiler/Instrumentation.cs
@@ -450,7 +450,8 @@ namespace Datadog.Trace.ClrProfiler
             Task.Run(
                 async () =>
                 {
-                    FakeRcm.Initialize();
+                    DynamicConfiguration.Initialize();
+
                     // TODO: LiveDebugger should be initialized in TracerManagerFactory so it can respond
                     // to changes in ExporterSettings etc.
